Fix deprecated np.float

+++ polyfuzz/models/_utils.py
@@ -88,7 +88,7 @@ def cosine_similarity(from_vector: np.ndarray,
 
         indices = _top_n_idx_sparse(similarity_matrix, top_n)
         similarities = _top_n_similarities_sparse(similarity_matrix, indices)
-        indices = np.array(np.nan_to_num(np.array(indices, dtype=np.float), nan=0), dtype=np.int)
+        indices = np.array(np.nan_to_num(np.array(indices, dtype=np.float32), nan=0), dtype=np.int32)
 
     # Faster than knn and slower than sparse but uses more memory
     else:
